Add docker image definition
Build three Docker image, one is for bugbug itself. It is just installing
bugbug and its dependencies.

One is for retrieving information from the mozilla-central Mercurial
repository, it depends on the first one and install the right Mercurial
version.

The last one is for retrieving information from the Bugzilla instance, it
depends in the first one and needs a valid Bugzilla token.