Import retriever services and add Docker image definition (#251)
* Import Retriever class from release-services repository

This basically import the `retriever.py` file from the `release-services`
repository at hash 77cdddd. I removed imports and reference to cli-common
helpers that will likely needs to be reimplemented, like the raven support.

The next commit will defines some Dockerfiles that will use the imported file.

* Add docker image definition

Build three Docker image, one is for bugbug itself. It is just installing
bugbug and its dependencies.

One is for retrieving information from the mozilla-central Mercurial
repository, it depends on the first one and install the right Mercurial
version.

The last one is for retrieving information from the Bugzilla instance, it
depends in the first one and needs a valid Bugzilla token.

* Separate the two tasks into separate script files

They share almost no code at all so they don't need to be in the same file

* Apply Black on the scripts to makes Flake8 happy