Create the data pipeline task definition
There is a hook (which runs every day at midnight) that will spawn the
data-pipeline using the latest docker images that were build on latest
release. The hook itself is updated on each release and is versionned in this
repository.

Add task for building the Docker image for HTTP service