Add the possibility for models to declare static extra data
This data will be saved and returned to the HTTP service customers.

There are other ways of exposing it, like having a dedicated endpoint for it
but it raises some questions about the web workers loading models and the
immediate goal is to remove the bugbug dependency from relman-auto-nag
project.