Add basic linear regression to detect the trend of metrics over time. #852
Internet suggested that using a linear regression on a timeseries data is a
first good step to detect the trend on the data. It seems to works fine but I
wasn't unable to distinguish between the metric getting significantly worse or
just some noise.

Because the range of value is limited, the slope is also very small.

I updated the checking script to emit a warming message instead of failing so
we can tune it if needing before activating it.