Make train/test split configurable by subclasses of Model
Some models might want to implement a different way to split between
train and test sets. For example, for test scheduling we'll need to
use the past past as the training set, and the recent past as the
test set. We can't mix them or the evaluation will be falsified.