Split regressor finder task in four separate tasks
A task to find commits to ignore, a task to classify commits between
bug-fixing vs not-bug-fixing, a task to find regressors using the
normal repo, and a task to find regressors using the tokenized
repo.

This way we can also find regressors for the two kinds of repos
in parallel.

Fixes #1273

Make the past bugs by function task depend on the task to classify
commits between bug-fixing and not-bug-fixing.

Fixes #1274