Consider Objective-C/C++ files differently than normal C/C++ files
For Objective-C/C++ files, we use both the "C/C++" and "Objective-C/C++" type
as they are few but share most characteristics with C/C++ files. We don't want
to lose this information by just overwriting the type, but we want the
additional information that it is an Objective-C/C++ file.

Fixes #1221