Convert IdleTTLCache to a ReadthroughTTLCache
Following PR feedback, converting IdleTTLCache to a data structure that can
handle all of our caching logic.  ReadthroughTTLCache is passed a callback, and
returns the value of the callback called on a key when an object is retrieved
from it. This value is stored for TTL seconds if it has been accessed within
the last TTL seconds.