Make the custom evaluation of the TestSelect models as close as possible to reality
First of all, use all possible schedulable tasks.

Then, make sure the results don't look better than they actually are by not leaking past
failure data from the training set to the test set by using for any given push the past
failure data info from 100 pushes before.
We can't use the past failure data for the push right before the push of interest because
we store past failure data in batches of 100 pushes.
In order to support this, we have to increase the history timespan
of the past failures data (at least until we find a nicer solution).

There is still some data leaking from training to test, but it should be pretty limited:
the touched together data and the failing together data is not snapshotted to the time of
the push of interest.