Bug 1622336 - Initialize schedules service when Hg push happens
This change adds a Pulse listener for Hg pushes as a background process
to `bugbug-http-service`.

This listener works if `PULSE_USER` and `PULSE_PASSWORD` are set before starting
up the Docker containers.

The listener only pays attention to autoland and try pushes. When these happen
we do a GET request from the bugbug deployment. This will get the schedules service
ready for when the Gecko decision task will query it.