Pull running 'hg pull' instead of using hglib to benefit from Mercurial's default transaction rollback
Currently we use hglib, so when we send SIGTERM to the process which is doing the pull, the process is terminated
without error handling and the repository is left in a bad state.
Instead, by using 'hg pull', we benefit from Mercurial's handling of SIGTERM: transaction rollback.

Might help with #1673