Add an option to the regressor model to use the regressor finder results only to exclude clean patches
With this option, the model is not going to use the regressor finder results as examples of
bug-introducing patches, but it will not use them as examples of clean patches either.

Since the regression finder results are often imprecise, this allows us to reduce the number
of patches we wrongly consider as bug-introducing, without increasing the number of patches
which we wrongly consider as not bug-introducing.