Use SendGrid to send quality notification
This way we can get rid of the ugly "Taskcluster" header in the emails, and we can simplify the markdown
by removing actually unneeded "<br>"s (in the future we might even just switch to plain HTML instead).

Also, we can now send it to one manager and set the other receivers in CC, fixing #2606.