Docs: Typo Fix (#4570)
Corrected "The aim of this classifier is to detect bugs which should be documented for developers." to "The aim of this classifier is to detect bugs that should be documented for developers.

This pull request addresses a minor typo found in repository. The typo has been corrected to improve clarity and maintain the quality of the documentation.

This change is purely cosmetic and does not affect functionality.