Use "default" instead of "tip" for Mercurial
"tip" is an ambiguous reference, it can be the current tip of any branch in the repository. In many cases on hg.mozilla.org, this ends up being a tags branch.

"default" is safer as it will always be the tip of the default branch.

This will maybe help with #5239